Migrate create-project
